+++ examples/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @syllst/examples
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @syllst/processor@0.6.0
+
 ## 0.4.5
 
 ### Patch Changes