+++ packages/validator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @syllst/validator
 
+## 0.4.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @syllst/processor@0.6.0
+
 ## 0.4.6
 
 ### Patch Changes